move local-MP comm to its own module instead of cramping Platform.cpp
