+++ src/frontend/qt_sdl/CMakeLists.txt
@@ -25,6 +25,7 @@ set(SOURCES_QT_SDL
     Input.cpp
     LAN_PCap.cpp
     LAN_Socket.cpp
+	LocalMP.cpp
     OSD.cpp
     OSD_shaders.h
     font.h
